Create customizable Onboarding component
